feat: Add llm_output_type field to LLMTaskWorker
This commit adds a new field, llm_output_type, to the LLMTaskWorker class. This field allows specifying the output type of the LLM if it differs from the task output type. If llm_output_type is not set, the convention is to pick the first output type from the list of output types.