feat: Add support for sink workers in Graph class
This commit adds a new method `set_sink()` to the `Graph` class, which allows setting a worker as a data sink in the task graph. The sink worker consumes the output of the specified worker and the output can be retrieved after running the graph using the `get_tasks()` method.