refactor: Add output_type parameter to set_sink method in Graph class
The set_sink method in the Graph class now accepts an additional parameter, output_type, which specifies the type of task that the sink worker should consume. This ensures that the sink worker is compatible with the specified output type. The method throws a ValueError if the worker does not have the specified output type.