refactor: Update user request tiles dynamically
This commit updates the `updateUserRequests` function in the `index.html` file. It introduces dynamic updating of user request tiles based on the provided `userRequests` data. The existing tiles are now mapped by `data-task-id` and new request tiles are added if they don't already exist. Additionally, tiles that are no longer present in `userRequests` are removed.