feat: allow a TaskWorker to have multiple consumers expected the same task type
this requires that publish_work() needs to be called with the right consumer in case there is a choice.

i don't know yet whether this should be done by name or per object. for the cache we need to query by name, so perhaps the whole interface needs to be just by name as well.