fix: an issue where cached results specified a no-longer used consumer
a recent change in PlanAI allowed TaskWorkers to have multiple downstream consumers for the same task type. to correctly route cached tasks, the consumer was specified as part of the cached results. when rewiring a graph, this could lead to a disconnect. to fix this, we ignore the downstream consumer name when a taskworker only has one downstream consumer for the task